+++ components/divider/Divider.tsx
@@ -1,0 +1,50 @@
+import * as React from 'react';
+import classNames from 'classnames';
+import { ConfigContext } from '../config-provider';
+
+export interface DividerProps {
+  prefixCls?: string;
+  type?: 'horizontal' | 'vertical';
+  orientation?: 'left' | 'right' | 'center';
+  className?: string;
+  children?: React.ReactNode;
+  dashed?: boolean;
+  style?: React.CSSProperties;
+  plain?: boolean;
+}
+
+const Divider: React.FC<DividerProps> = props => {
+  const { getPrefixCls, direction } = React.useContext(ConfigContext);
+  const {
+    prefixCls: customizePrefixCls,
+    type = 'horizontal',
+    orientation = 'center',
+    className,
+    children,
+    dashed,
+    plain,
+    ...restProps
+  } = props;
+  const prefixCls = getPrefixCls('divider', customizePrefixCls);
+  const orientationPrefix = orientation.length > 0 ? `-${orientation}` : orientation;
+  const hasChildren = !!children;
+  const classString = classNames(
+    prefixCls,
+    `${prefixCls}-${type}`,
+    {
+      [`${prefixCls}-with-text`]: hasChildren,
+      [`${prefixCls}-with-text${orientationPrefix}`]: hasChildren,
+      [`${prefixCls}-dashed`]: !!dashed,
+      [`${prefixCls}-plain`]: !!plain,
+      [`${prefixCls}-rtl`]: direction === 'rtl',
+    },
+    className,
+  );
+  return (
+    <div className={classString} {...restProps} role="separator">
+      {children && <span className={`${prefixCls}-inner-text`}>{children}</span>}
+    </div>
+  );
+};
+
+export default Divider;
