+++ components/divider/demo/group.md
@@ -1,0 +1,27 @@
+---
+order: 99
+title:
+  zh-CN: 垂直分割线组
+  en-US: Vertical Group
+---
+
+## zh-CN
+
+使用 `Divider.Group` 设置为行内的垂直分割线。
+
+## en-US
+
+Use `Divider.Group` make it vertical.
+
+```jsx
+import { Divider } from 'antd';
+
+ReactDOM.render(
+  <Divider.Group>
+    Text
+    <a href="#">Link</a>
+    <a href="#">Link</a>
+  </Divider.Group>,
+  mountNode,
+);
+```
