+++ components/divider/index.zh-CN.md
@@ -15,6 +15,8 @@ cover: https://gw.alipayobjects.com/zos/alicdn/5swjECahe/Divider.svg
 
 ## API
 
+### Divider
+
 | 参数        | 说明                       | 类型                          | 默认值       | 版本  |
 | ----------- | -------------------------- | ----------------------------- | ------------ | ----- |
 | className   | 分割线样式类               | string                        | -            |       |
@@ -23,3 +25,10 @@ cover: https://gw.alipayobjects.com/zos/alicdn/5swjECahe/Divider.svg
 | plain       | 文字是否显示为普通正文样式 | boolean                       | false        | 4.2.0 |
 | style       | 分割线样式对象             | CSSProperties                 | -            |       |
 | type        | 水平还是垂直类型           | `horizontal` \| `vertical`    | `horizontal` |       |
+
+### Divider.Group
+
+| 参数   | 说明             | 类型                       | 默认值     | 版本  |
+| ------ | ---------------- | -------------------------- | ---------- | ----- |
+| dashed | 是否虚线         | boolean                    | false      | 4.7.0 |
+| type   | 水平还是垂直类型 | `horizontal` \| `vertical` | `vertical` | 4.7.0 |
